Move characteristic form to reusable component and create edit page
